requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.0
requests==2.32.3
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
zipp==3.21.0
